+++ crates/node/src/service.rs
@@ -1,15 +1,16 @@
 //! Service and ServiceFactory implementation. Specialized wrapper over substrate service.
 
 use std::cell::RefCell;
-use std::marker::PhantomData;
 use std::path::PathBuf;
 use std::sync::Arc;
 use std::time::Duration;
 
 use futures::channel::mpsc;
 use futures::future;
+use futures::future::BoxFuture;
 use futures::prelude::*;
-use madara_runtime::{self, Hash, RuntimeApi, StarknetHasher};
+use madara_runtime::opaque::Block;
+use madara_runtime::{self, Hash, RuntimeApi, SealingMode, StarknetHasher};
 use mc_block_proposer::ProposerFactory;
 use mc_data_availability::avail::config::AvailConfig;
 use mc_data_availability::avail::AvailClient;
@@ -24,12 +25,10 @@ use mc_transaction_pool::FullPool;
 use mp_sequencer_address::{
     InherentDataProvider as SeqAddrInherentDataProvider, DEFAULT_SEQUENCER_ADDRESS, SEQ_ADDR_STORAGE_KEY,
 };
-use madara_runtime::opaque::Block;
 use parity_scale_codec::Encode;
 use prometheus_endpoint::Registry;
 use sc_client_api::{Backend, BlockBackend, BlockchainEvents, HeaderBackend};
-use sc_consensus::BasicQueue;
-use sc_consensus::BlockImportParams;
+use sc_consensus::{BasicQueue, BlockImportParams};
 use sc_consensus_aura::{SlotProportion, StartAuraParams};
 use sc_consensus_grandpa::{GrandpaBlockImport, SharedVoterState};
 use sc_consensus_manual_seal::{ConsensusDataProvider, Error};
@@ -37,21 +36,15 @@ pub use sc_executor::NativeElseWasmExecutor;
 use sc_service::error::Error as ServiceError;
 use sc_service::{new_db_backend, Configuration, TaskManager, WarpSyncParams};
 use sc_telemetry::{Telemetry, TelemetryHandle, TelemetryWorker};
-use sp_api::ProvideRuntimeApi;
 use sp_api::offchain::OffchainStorage;
-use sp_api::{ConstructRuntimeApi, TransactionFor};
+use sp_api::{ConstructRuntimeApi, ProvideRuntimeApi, TransactionFor};
 use sp_consensus_aura::sr25519::AuthorityPair as AuraPair;
 use sp_inherents::InherentData;
 use sp_offchain::STORAGE_PREFIX;
-use sp_runtime::testing::Digest;
-use sp_runtime::testing::DigestItem;
-use sp_runtime::traits::BlakeTwo256;
-use sp_runtime::traits::Block as BlockT;
+use sp_runtime::testing::{Digest, DigestItem};
+use sp_runtime::traits::{BlakeTwo256, Block as BlockT};
 use sp_trie::PrefixedMemoryDB;
-use mc_deoxys::{fetch_block, BlockQueue, create_block_queue};
-use lazy_static::lazy_static;
 
-use crate::commands::Sealing;
 use crate::genesis_block::MadaraGenesisBlockBuilder;
 use crate::rpc::StarknetDeps;
 use crate::starknet::{db_config_dir, MadaraBackend};
@@ -86,6 +79,7 @@ type BoxBlockImport<Client> = sc_consensus::BoxBlockImport<Block, TransactionFor
 pub fn new_partial<BIQ>(
     config: &Configuration,
     build_import_queue: BIQ,
+    cache_more_things: bool,
 ) -> Result<
     sc_service::PartialComponents<
         FullClient,
@@ -174,7 +168,7 @@ where
         telemetry.as_ref().map(|x| x.handle()),
     )?;
 
-    let madara_backend = Arc::new(MadaraBackend::open(&config.database, &db_config_dir(config))?);
+    let madara_backend = Arc::new(MadaraBackend::open(&config.database, &db_config_dir(config), cache_more_things)?);
 
     let (import_queue, block_import) = build_import_queue(
         client.clone(),
@@ -259,23 +253,23 @@ where
         ),
         Box::new(client),
     ))
-
-}
-
-lazy_static! {
-    static ref QUEUE: BlockQueue = create_block_queue();
 }
 
 /// Builds a new service for a full client.
-pub async fn new_full(
+///
+/// # Arguments
+///
+/// - `cache`: whether more information should be cached when storing the block in the database.
+pub fn new_full(
     config: Configuration,
-    sealing: Option<Sealing>,
+    sealing: SealingMode,
     da_layer: Option<(DaLayer, PathBuf)>,
     rpc_port: u16,
+    cache_more_things: bool,
     network_uri: &str,
 ) -> Result<TaskManager, ServiceError> {
     let build_import_queue =
-        if sealing.is_some() { build_manual_seal_import_queue } else { build_aura_grandpa_import_queue };
+        if sealing.is_default() { build_aura_grandpa_import_queue } else { build_manual_seal_import_queue };
 
     let sc_service::PartialComponents {
         client,
@@ -286,7 +280,7 @@ pub async fn new_full(
         select_chain,
         transaction_pool,
         other: (block_import, grandpa_link, mut telemetry, madara_backend),
-    } = new_partial(&config, build_import_queue)?;
+    } = new_partial(&config, build_import_queue, cache_more_things)?;
 
     let mut net_config = sc_network::config::FullNetworkConfiguration::new(&config.network);
 
@@ -295,9 +289,7 @@ pub async fn new_full(
         &config.chain_spec,
     );
 
-    let warp_sync_params = if sealing.is_some() {
-        None
-    } else {
+    let warp_sync_params = if sealing.is_default() {
         net_config
             .add_notification_protocol(sc_consensus_grandpa::grandpa_peers_set_config(grandpa_protocol_name.clone()));
         let warp_sync = Arc::new(sc_consensus_grandpa::warp_proof::NetworkProvider::new(
@@ -306,6 +298,8 @@ pub async fn new_full(
             Vec::default(),
         ));
         Some(WarpSyncParams::WithProvider(warp_sync))
+    } else {
+        None
     };
 
     let (network, system_rpc_tx, tx_handler_controller, network_starter, sync_service) =
@@ -328,14 +322,18 @@ pub async fn new_full(
     let force_authoring = config.force_authoring;
     let backoff_authoring_blocks: Option<()> = None;
     let name = config.network.node_name.clone();
-    let enable_grandpa = !config.disable_grandpa && sealing.is_none();
+    let enable_grandpa = !config.disable_grandpa && sealing.is_default();
     let prometheus_registry = config.prometheus_registry().cloned();
     let starting_block = client.info().best_number;
 
     // Channel for the rpc handler to communicate with the authorship task.
-    // TODO: commands_stream is is currently unused, but should be used to implement the `sealing`
-    // parameter
-    let (command_sink, commands_stream) = mpsc::channel(1000);
+    let (command_sink, commands_stream) = match sealing {
+        SealingMode::Manual => {
+            let (sender, receiver) = mpsc::channel(1000);
+            (Some(sender), Some(receiver))
+        }
+        _ => (None, None),
+    };
 
     let overrides = overrides_handle(client.clone());
     let starknet_rpc_params = StarknetDeps {
@@ -358,7 +356,7 @@ pub async fn new_full(
                 graph: graph.clone(),
                 deny_unsafe,
                 starknet: starknet_rpc_params.clone(),
-                command_sink: if sealing.is_some() { Some(command_sink.clone()) } else { None },
+                command_sink: command_sink.clone(),
             };
             crate::rpc::create_full(deps).map_err(Into::into)
         })
@@ -382,7 +380,7 @@ pub async fn new_full(
     task_manager.spawn_essential_handle().spawn(
         "mc-mapping-sync-worker",
         Some("madara"),
-        MappingSyncWorker::new(
+        MappingSyncWorker::<_, _, _, StarknetHasher>::new(
             client.import_notification_stream(),
             Duration::new(6, 0),
             client.clone(),
@@ -390,7 +388,6 @@ pub async fn new_full(
             madara_backend.clone(),
             3,
             0,
-            PhantomData::<StarknetHasher>,
         )
         .for_each(|()| future::ready(())),
     );
@@ -426,8 +423,14 @@ pub async fn new_full(
 
     if role.is_authority() {
         // manual-seal authorship
-        if let Some(sealing) = sealing {
+        if !sealing.is_default() {
+            // NOTE(nils-mathieu):
+            //   For now I used a channel of size 20, this should be plently enough. That might
+            //   become a config option in the future.
+            let (block_sender, block_receiver) = async_channel::bounded::<mp_block::Block>(20);
+
             run_manual_seal_authorship(
+                block_receiver,
                 sealing,
                 client,
                 transaction_pool,
@@ -442,7 +445,7 @@ pub async fn new_full(
 
             let network_uri = Box::from(network_uri);
             tokio::spawn(async move {
-                fetch_block(QUEUE.clone(), &network_uri, rpc_port).await;
+                mc_deoxys::fetch_block(block_sender, &network_uri, rpc_port).await;
             });
 
             log::info!("Manual Seal Ready");
@@ -550,20 +553,21 @@ pub async fn new_full(
     }
 
     network_starter.start_network();
-    
+
     Ok(task_manager)
 }
 
 #[allow(clippy::too_many_arguments)]
 fn run_manual_seal_authorship(
-    sealing: Sealing,
+    block_receiver: async_channel::Receiver<mp_block::Block>,
+    sealing: SealingMode,
     client: Arc<FullClient>,
     transaction_pool: Arc<FullPool<Block, FullClient>>,
     select_chain: FullSelectChain,
     block_import: BoxBlockImport<FullClient>,
     task_manager: &TaskManager,
     prometheus_registry: Option<&Registry>,
-    commands_stream: mpsc::Receiver<sc_consensus_manual_seal::rpc::EngineCommand<Hash>>,
+    commands_stream: Option<mpsc::Receiver<sc_consensus_manual_seal::rpc::EngineCommand<Hash>>>,
 ) -> Result<(), ServiceError>
 where
     RuntimeApi: ConstructRuntimeApi<Block, FullClient>,
@@ -610,51 +614,58 @@ where
     };
 
     struct QueryBlockConsensusDataProvider<C> {
-		_client: Arc<C>,
-	}
+        _client: Arc<C>,
 
-	impl<B, C> ConsensusDataProvider<B> for QueryBlockConsensusDataProvider<C>
-		where
-		B: BlockT,
-		C: ProvideRuntimeApi<B> + Send + Sync,{
-		type Transaction = TransactionFor<C, B>;
-		type Proof = ();
+        /// The receiver that we're using to receive blocks.
+        block_receiver: async_channel::Receiver<mp_block::Block>,
+    }
 
-		fn create_digest(&self, _parent: &B::Header, _inherents: &InherentData) -> Result<Digest, Error> {
-            let mut queue_guard = QUEUE.lock().unwrap();
-            let starknet_block: mp_block::Block = queue_guard.pop_front().unwrap();
+    impl<B, C> ConsensusDataProvider<B> for QueryBlockConsensusDataProvider<C>
+    where
+        B: BlockT,
+        C: ProvideRuntimeApi<B> + Send + Sync,
+    {
+        type Transaction = TransactionFor<C, B>;
+        type Proof = ();
 
-            let block_digest_item: DigestItem = sp_runtime::DigestItem::PreRuntime(mp_digest_log::MADARA_ENGINE_ID, Encode::encode(&starknet_block));
+        fn create_digest(&self, _parent: &B::Header, _inherents: &InherentData) -> Result<Digest, Error> {
+            let starknet_block: mp_block::Block = self.block_receiver.recv_blocking().unwrap();
+
+            let block_digest_item: DigestItem =
+                sp_runtime::DigestItem::PreRuntime(mp_digest_log::MADARA_ENGINE_ID, Encode::encode(&starknet_block));
             Ok(Digest { logs: vec![block_digest_item] })
         }
 
-		fn append_block_import(
-			&self,
-			_parent: &B::Header,
-			params: &mut BlockImportParams<B, Self::Transaction>,
-			_inherents: &InherentData,
-			_proof: Self::Proof,
-		) -> Result<(), Error> {
-			params.post_digests.push(DigestItem::Other(vec![1]));
-			Ok(())
-		}
-	}
+        fn append_block_import(
+            &self,
+            _parent: &B::Header,
+            params: &mut BlockImportParams<B, Self::Transaction>,
+            _inherents: &InherentData,
+            _proof: Self::Proof,
+        ) -> Result<(), Error> {
+            params.post_digests.push(DigestItem::Other(vec![1]));
+            Ok(())
+        }
+    }
 
-    let manual_seal = match sealing {
-        Sealing::Manual => future::Either::Left(sc_consensus_manual_seal::run_manual_seal(
-            sc_consensus_manual_seal::ManualSealParams {
+    let manual_seal: BoxFuture<_> = match sealing {
+        SealingMode::Manual => {
+            Box::pin(sc_consensus_manual_seal::run_manual_seal(sc_consensus_manual_seal::ManualSealParams {
                 block_import,
                 env: proposer_factory,
                 client: client.clone(),
                 pool: transaction_pool,
-                commands_stream,
+                commands_stream: commands_stream.expect("Manual sealing requires a channel from RPC."),
                 select_chain,
-                consensus_data_provider: Some(Box::new(QueryBlockConsensusDataProvider { _client: client})),
+                consensus_data_provider: Some(Box::new(QueryBlockConsensusDataProvider {
+                    _client: client,
+                    block_receiver,
+                })),
                 create_inherent_data_providers,
-            },
-        )),
-        Sealing::Instant => future::Either::Right(sc_consensus_manual_seal::run_instant_seal(
-            sc_consensus_manual_seal::InstantSealParams {
+            }))
+        }
+        SealingMode::Instant { finalize } => {
+            let instant_seal_params = sc_consensus_manual_seal::InstantSealParams {
                 block_import,
                 env: proposer_factory,
                 client,
@@ -662,15 +673,20 @@ where
                 select_chain,
                 consensus_data_provider: None,
                 create_inherent_data_providers,
-            },
-        )),
+            };
+            if finalize {
+                Box::pin(sc_consensus_manual_seal::run_instant_seal_and_finalize(instant_seal_params))
+            } else {
+                Box::pin(sc_consensus_manual_seal::run_instant_seal(instant_seal_params))
+            }
+        }
+        _ => unreachable!("Other sealing modes are not expected in manual-seal."),
     };
 
     // we spawn the future on a background thread managed by service.
     task_manager.spawn_essential_handle().spawn_blocking("manual-seal", None, manual_seal);
     Ok(())
 }
-
 
 type ChainOpsResult = Result<
     (
@@ -683,9 +699,9 @@ type ChainOpsResult = Result<
     ServiceError,
 >;
 
-pub fn new_chain_ops(config: &mut Configuration) -> ChainOpsResult {
+pub fn new_chain_ops(config: &mut Configuration, cache_more_things: bool) -> ChainOpsResult {
     config.keystore = sc_service::config::KeystoreConfig::InMemory;
     let sc_service::PartialComponents { client, backend, import_queue, task_manager, other, .. } =
-        new_partial::<_>(config, build_aura_grandpa_import_queue)?;
+        new_partial::<_>(config, build_aura_grandpa_import_queue, cache_more_things)?;
     Ok((client, backend, import_queue, task_manager, other.3))
 }