+++ crates/primitives/transactions/src/conversions.rs
@@ -1,4 +1,5 @@
 use alloc::sync::Arc;
+use alloc::vec::Vec;
 
 use blockifier::execution::contract_class::ContractClass;
 use blockifier::transaction::objects::TransactionExecutionResult;
@@ -9,15 +10,13 @@ use starknet_api::api_core::Nonce;
 use starknet_api::hash::StarkFelt;
 use starknet_api::transaction as sttx;
 use starknet_api::transaction::{Fee, TransactionVersion};
-use alloc::vec::Vec;
-
-use crate::DeployTransaction;
 
 use super::compute_hash::ComputeTransactionHash;
 use super::{
     DeclareTransaction, DeclareTransactionV0, DeclareTransactionV1, DeclareTransactionV2, DeployAccountTransaction,
     HandleL1MessageTransaction, InvokeTransaction, InvokeTransactionV0, InvokeTransactionV1,
 };
+use crate::DeployTransaction;
 
 impl DeclareTransactionV0 {
     fn try_into_executable<H: HasherT>(
@@ -157,7 +156,7 @@ impl InvokeTransactionV0 {
             signature: inner.signature.0.iter().map(|felt| Felt252Wrapper::from(*felt)).collect(),
             contract_address: inner.contract_address.into(),
             entry_point_selector: inner.entry_point_selector.into(),
-            calldata: inner.calldata.0.iter().map(|felt| Felt252Wrapper::from(*felt)).collect()
+            calldata: inner.calldata.0.iter().map(|felt| Felt252Wrapper::from(*felt)).collect(),
         }
     }
 }
@@ -184,7 +183,7 @@ impl InvokeTransactionV1 {
             signature: inner.signature.0.iter().map(|felt| Felt252Wrapper::from(*felt)).collect(),
             nonce: inner.nonce.into(),
             sender_address: inner.sender_address.into(),
-            calldata: inner.calldata.0.iter().map(|felt| Felt252Wrapper::from(*felt)).collect()
+            calldata: inner.calldata.0.iter().map(|felt| Felt252Wrapper::from(*felt)).collect(),
         }
     }
 }
@@ -231,7 +230,7 @@ impl DeployAccountTransaction {
             nonce: inner.nonce.into(),
             contract_address_salt: inner.contract_address_salt.into(),
             constructor_calldata: inner.constructor_calldata.0.iter().map(|felt| Felt252Wrapper::from(*felt)).collect(),
-            class_hash: inner.class_hash.into()
+            class_hash: inner.class_hash.into(),
         }
     }
 }
@@ -271,7 +270,7 @@ impl HandleL1MessageTransaction {
         Self {
             nonce: u64::try_from(inner.nonce.0).unwrap(),
             contract_address: inner.contract_address.into(),
-            entry_point_selector:  inner.entry_point_selector.0.into(),
+            entry_point_selector: inner.entry_point_selector.0.into(),
             calldata: inner.calldata.0.iter().map(|felt| Felt252Wrapper::from(*felt)).collect(),
         }
     }
@@ -288,7 +287,7 @@ fn vec_of_felt_to_calldata(felts: &[Felt252Wrapper]) -> sttx::Calldata {
 impl From<sttx::DeclareTransactionV0V1> for DeclareTransactionV0 {
     fn from(remote: sttx::DeclareTransactionV0V1) -> Self {
         DeclareTransactionV0 {
-            max_fee: remote.max_fee.0,  // This assumes the `Fee` type's internal value can be accessed this way.
+            max_fee: remote.max_fee.0, // This assumes the `Fee` type's internal value can be accessed this way.
             signature: signature_to_vec_of_felt(&remote.signature),
             nonce: remote.nonce.into(),
             class_hash: remote.class_hash.into(),
@@ -301,7 +300,7 @@ impl From<sttx::DeclareTransactionV0V1> for DeclareTransactionV0 {
 impl From<sttx::DeclareTransactionV0V1> for DeclareTransactionV1 {
     fn from(remote: sttx::DeclareTransactionV0V1) -> Self {
         DeclareTransactionV1 {
-            max_fee: remote.max_fee.0,  // Assuming `Fee` type's inner value can be accessed with `.0`.
+            max_fee: remote.max_fee.0, // Assuming `Fee` type's inner value can be accessed with `.0`.
             signature: signature_to_vec_of_felt(&remote.signature),
             nonce: remote.nonce.into(),
             class_hash: remote.class_hash.into(),
@@ -314,7 +313,7 @@ impl From<sttx::DeclareTransactionV0V1> for DeclareTransactionV1 {
 impl From<sttx::DeclareTransactionV2> for DeclareTransactionV2 {
     fn from(remote: sttx::DeclareTransactionV2) -> Self {
         DeclareTransactionV2 {
-            max_fee: remote.max_fee.0,  // Assuming `Fee` type's inner value can be accessed with `.0`.
+            max_fee: remote.max_fee.0, // Assuming `Fee` type's inner value can be accessed with `.0`.
             signature: signature_to_vec_of_felt(&remote.signature),
             nonce: remote.nonce.into(),
             class_hash: remote.class_hash.into(),